third_party/WebKit/LayoutTests/fast/css/first-letter-hover-hit-test.html

<!doctype html>
<script src="../../resources/js-test.js"></script>
<style>
#tab:first-letter {
  text-transform: uppercase;
}
#tab {
  display: block;
}
a:hover {
  color: orange;
}
</style>
<div>
  <span id="tab"><a id='a'>H</a></span>
</div>

<script>
if (window.eventSender) {
  document.body.offsetTop; // Force layout. The mouse is not tracked before first layout.

  offset = document.getElementById('tab').getBoundingClientRect();
  eventSender.mouseMoveTo(offset.left + 5, offset.top + 5);

  document.body.offsetTop; // Update layout for hovered state.

  style = window.getComputedStyle(document.getElementById('a'));
  shouldBe('style.color', "'rgb(255, 165, 0)'");
} else {
  debug("This test requires DRT.");
}
</script>
