third_party/blink/web_tests/platform/generic/fast/css/first-letter-hover-hit-test-expected.txt

PASS style.color is 'rgb(255, 165, 0)'
PASS successfullyParsed is true

TEST COMPLETE

H
